+++ frontend/src/app/shared/services/index.ts
@@ -3,3 +3,4 @@ export * from './message.service';
 export * from './openapi.service';
 export * from './storage.service';
 export * from './paginator-i18n.service';
+export * from './clipboard.service';